Update hybridDaughter.gdml: repair COIL_BOX solid
The <box> named "COIL_BOX" in the <solids> block in hybridDaughter.gdml was in conflict with the <box> named "COIL_BOX" in the <solids> block in upstreamDaughter.gdml. In order to repair this, the name of the <box> named "COIL_BOX" in the <solids> block in hybridDaughter.gdml has been changed to "COIL_BOX_1". Essentially, the double use of the name "COIL_BOX" in two different .gdml files seemed to have confused GEANT. Secondly, once the correct definition for "COIL_BOX_1" was used, its volume was not sufficiently large to have a non-trivially zero union with SEGMENT_X (they didn't overlap). "COIL_BOX_1" was enlarged by 2000 in the z direction without penalty, because it appears only once in the code, and only as a union, and it does not change the center point of the union. In other words, it only needs to have the correct center, and it must fully contain SEGMENT_X.